rename `gh subscription get` to `gh subscription view`
