+++ pkg/cmd/subscription/subscription.go
@@ -2,8 +2,8 @@ package subscription
 
 import (
 	"github.com/MakeNowJust/heredoc"
-	subGetCmd "github.com/cli/cli/v2/pkg/cmd/subscription/get"
 	subListCmd "github.com/cli/cli/v2/pkg/cmd/subscription/list"
+	subViewCmd "github.com/cli/cli/v2/pkg/cmd/subscription/view"
 	"github.com/cli/cli/v2/pkg/cmdutil"
 	"github.com/spf13/cobra"
 )
@@ -16,14 +16,14 @@ func NewCmdSubscription(f *cmdutil.Factory) *cobra.Command {
 		Aliases: []string{"sub"},
 		Example: heredoc.Doc(`
 			$ gh subscription list
-			$ gh subscription get monalisa/hello-world
+			$ gh subscription view monalisa/hello-world
 		`),
 		GroupID: "core",
 	}
 
 	cmdutil.AddGroup(cmd, "General Commands",
 		subListCmd.NewCmdList(f, nil),
-		subGetCmd.NewCmdGet(f, nil),
+		subViewCmd.NewCmdView(f, nil),
 	)
 
 	return cmd
